rename recipe folder inside component folder
